cypress component_spec test fix
